Merge duplicate nodes when inserting
Add a new constructor to BinaryTree's InsertResult: InsertMerge. This
constructor is used when a duplicate value is about to be inserted so it
should be merged instead.

RedBlackTree's insert now checks if the focused node of the zipper is
black to avoid unnecessary rotations. A black node means that the node
was merged rather that inserted so no further action is needed.

Update README.